requirements.txt
streamlit
sklearn
numpy
pandas
pillow